renaming getTokenRegistry >> getTokenInfo
