+++ src/lib/dvf/token/getTokenInfo.js
